requirements/base.txt
Django
psycopg2
django-allauth
Pillow
django-redis
django-mptt
django-froala-editor
dj-pagination
webstack-django-sorting
django-auditlog
jsonfield
lxml
ics
wikipedia
pytz
cssselect
djangorestframework
openpyxl
reportlab
django-solo
django-fcm
django-versatileimagefield
django-jet
django-admin-sortable2
django-filter